Move and update website.models
